+++ R/ycevo-package.R
@@ -1,5 +1,3 @@
-
-
 #' Nonparametric Estimation of the Yield Curve Evolution
 #'
 #' Nonparametric estimation of discount functions and yield curves.
@@ -28,13 +26,16 @@
 #' @importFrom ggplot2 aes
 #' @importFrom ggplot2 geom_line
 #' @importFrom ggplot2 ggplot
+#' @importFrom graphics legend
+#' @importFrom graphics lines
+#' @importFrom graphics par
 #' @importFrom lubridate ymd
 #' @importFrom Matrix colSums rowSums sparseMatrix t
 #' @importFrom Rcpp evalCpp
 #' @importFrom Rcpp sourceCpp
-#' @importFrom rlang :=
 #' @importFrom rlang !! sym .data
 #' @importFrom rlang %||%
+#' @importFrom rlang :=
 #' @importFrom rlang syms
 #' @importFrom stats runif
 #' @importFrom tibble tibble
@@ -43,7 +44,6 @@
 #' @useDynLib ycevo
 #' @references Koo, B., La Vecchia, D., & Linton, O. (2021). Estimation of a nonparametric model for bond prices from cross-section and time series information. Journal of Econometrics, 220(2), 562-588.
 "_PACKAGE"
-
 
 
 utils::globalVariables(c("."))